A lot has happened here.
Added files for dealing with picbreeder schema, win-setup, and for adding functions to cppn factory.

Haven't verified that everything generates properly, but it appears to be creating neat genomes from a seed object. This is a big step. Soon, saving will be possible, then this tangle of objects can be figured out more completely.

What's the best architecture for something like this? Is this an easy to understand example? What would make it easier to understand? Easier to use?

Should more things be encompassed in a single module to assign a better purpose to each module? Other than small little thing x and small little thing y with 1 million modules.