Many changes here. One of the really nice changes is a phylogeny creator. It's pretty slick, but right now is integrated into just this experiment. In the future, I'd like to make it a separate component.
I am tired. Mind, body, soul. Finished the picbreeder image for the paper, backed up the artifacts. Next step: iesor validation once and for all. Then, once validated, jerry rig it into the iec framework. I want to do an infinite scroll thing (where you scroll down and then evolution starts running for a while, and loads up new individuals when it's done).

Then you examine the individuals (one at a time or something), then choose the next individual to be the seed. This seed moves to the top of the page, and the process starts over.